Move L6470 code to 'libs' folder
