+++ Marlin/src/HAL/shared/HAL_spi_L6470.cpp
@@ -44,7 +44,7 @@
 // --------------------------------------------------------------------------
 
 #include "../../core/serial.h"
-#include "../../module/L6470/L6470_Marlin.h"
+#include "../../libs/L6470/L6470_Marlin.h"
 
 // --------------------------------------------------------------------------
 // Software L6470 SPI
