+++ Marlin/src/Marlin.cpp
@@ -161,7 +161,7 @@
 #endif
 
 #if HAS_DRIVER(L6470)
-  #include "module/L6470/L6470_Marlin.h"
+  #include "libs/L6470/L6470_Marlin.h"
 #endif
 
 bool Running = true;
