+++ Marlin/src/Marlin.h
@@ -32,7 +32,7 @@
 #include <stdlib.h>
 
 #if HAS_DRIVER(L6470)
-  #include "module/L6470/L6470_Marlin.h"
+  #include "libs/L6470/L6470_Marlin.h"
   extern uint8_t axis_known_position;
 #endif
